Starts please_Find_Lagrangian_Forces_On_Eulerian...
This function has many dependencies and will go in a file by itself.

Ran into trouble inside of first dependency - can't find a function
implementation - so this commit goes up to there for now. However, there
is major debugging to be done concerning numpy arrays. Pure assignment is
by reference in general, and column vectors are not the way to go in numpy
- calling into the array will return an array instead of a value,
resulting in a reference assignment. Anyway, you don't need a column to be
multiplied by a row in numpy - dot opreation takes care of it. So... all
assignments and np.zeros need to be re-looked at for this behavior.