Ports Rubberband.m, fixes small error in original
The .beam file was previously using an indexing scheme starting at 1
instead of 0, which throws an error in the Python port. Accordingly,
Rubberband.m has been ported and the beam and vertex files updated. The
vertex file is identical, but the beam file shows slight differences in
some of the data (not all), which is unexpected. I'm leaving that alone
for now. Matplotlib is used for plotting, but has been commented in this
file because it was crashing on Windows 10 outside of an Anaconda
environment.