Ports plotting routine, debugging of code needed
This commit finishes the first phase of the Python port. The code now runs
and plots, but the loop dies after a certain number of time-steps. This
commit includes some first debugging toward fixing this problem.
Additionally, the code is incredibly slow compared to the MATLAB version,
likely because of issues such as using column major instead of row major
indexing.