Fixes bugs: rubberband example runs (SLOWLY!!)
This commit fixes bugs in give_Fluid_Pressure (inside
please_Update_Fluid_Velocity.py) that caused blowup after a certain number
of time steps. The port now completely runs the rubberband example (albeit
very slowly), and appears to deliver a matching result to the MATLAB
version (this assessment just based on a quick look at one of the output
files). The next steps should be to improve the speed of the program,
probably starting by moving to a row-major format, and to port another
example or two to test certain functions that were skipped in the
rubberband example.