Reverts commit & optimizes Delta_Indices functions
It turns out that row vs. column major concerns, while making things 2-3x
slower in MATLAB, do not really matter for ndarrays. What does matter is
that looping over ndarrays is REALLY slow - 10x or more slower than
MATLAB. To do this efficiently, we need to replace all loops with numPy
array manipulation routines, which leverage C and are quite fast. In a
first step toward this effort, this commit reverts the changes in the last
couple of commits to the functions give_1D_NonZero_Delta_Indices and
give_NonZero_Delta_Indices_XY, instead performing the ndarray manipulation
with numPy. The arrays will remain the same internally as they were in
MATLAB for easy comparison.