Both Info lists -> dict, grid_Info keys meaningful
grid_Info and model_Info are now dicts, since this data structure exactly
matches their use. Not only does this provide for a (marginally in this
case) faster lookup of values, but it allows for human-meaninful key
values for more readable code. This commit implements such key values for
grid_Info; model_Info is still TODO. As another added benefit, one could
pickle these dictionaries or write them to text for storing with the
result, providing for automatic documentation.